\name{codonToAAone}
\alias{codonToAAone}
\title{Translating Codons to Amino Acids}
\usage{
  codonToAAone(codon)
}
\arguments{
  \item{codon}{a DNA 3-character codon}
}
\description{
  Translate a DNA 3-character codon to an amino acid (one
  character)
}
\examples{
codonToAAone('AAA')
}
\seealso{
  \code{\link{codonToAAthree}}
}

