\name{codonToAAoneRNA}
\alias{codonToAAoneRNA}
\title{Translating Codons to Amino Acids}
\usage{
  codonToAAoneRNA(codon)
}
\arguments{
  \item{codon}{a RNA 3-character codon}
}
\description{
  Translate a RNA 3-character codon to an amino acid (one
  character)
}
\examples{
codonToAAoneRNA('AAA')
}
\seealso{
  \code{\link{codonToAAthreeRNA}}
}

