\name{codonToAAthree}
\alias{codonToAAthree}
\title{Translating Codons to Amino Acids}
\usage{
  codonToAAthree(codon)
}
\arguments{
  \item{codon}{a DNA 3-character codon}
}
\description{
  Translate a DNA 3-character codon to an amino acid (three
  character)
}
\examples{
codonToAAthree('AAA')
}
\seealso{
  \code{\link{codonToAAone}}
}

