\name{codonToAAthreeRNA}
\alias{codonToAAthreeRNA}
\title{Translating Codons to Amino Acids}
\usage{
  codonToAAthreeRNA(codon)
}
\arguments{
  \item{codon}{a RNA 3-character codon}
}
\description{
  Translate a RNA 3-character codon to an amino acid (three
  character)
}
\examples{
codonToAAthreeRNA('AAA')
}
\seealso{
  \code{\link{codonToAAoneRNA}}
}

